⚡ Bolt: Optimize Cache and fix CI by breaking circular dependencies
- Optimized Cache to O(1) LRU and amortized O(1) TTL cleanup.
- Moved EnvLoader to a dedicated file to break circular dependencies between environment and cache configs.
- Hardened EnvLoader against non-Node.js runtimes (Edge/Workers).
- Cleaned up src/lib/config/index.ts to remove triple and redundant exports.
- Switched Cache utility to direct configuration imports to avoid index module side-effects in Edge runtime.

Verified with benchmark scripts showing ~90% reduction in set() latency.
